Moves functions to lib.rs
